# shared-contracts/openapi/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kairos Core Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/knowledge/domains": {
      "post": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "Create Knowledge Domain",
        "operationId": "create_knowledge_domain_v1_knowledge_domains_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDomainCreatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Knowledge Domain V1 Knowledge Domains Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "List Knowledge Domains",
        "operationId": "list_knowledge_domains_v1_knowledge_domains_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Knowledge Domains V1 Knowledge Domains Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/knowledge/nodes": {
      "post": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "Create Knowledge Node",
        "operationId": "create_knowledge_node_v1_knowledge_nodes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNodeCreatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Knowledge Node V1 Knowledge Nodes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "List Knowledge Nodes",
        "operationId": "list_knowledge_nodes_v1_knowledge_nodes_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Org Id"
            }
          },
          {
            "name": "domain_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Knowledge Nodes V1 Knowledge Nodes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/knowledge/edges": {
      "post": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "Create Knowledge Edge",
        "operationId": "create_knowledge_edge_v1_knowledge_edges_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEdgeCreatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Knowledge Edge V1 Knowledge Edges Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-index"
        ],
        "summary": "List Knowledge Edges",
        "operationId": "list_knowledge_edges_v1_knowledge_edges_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Org Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Knowledge Edges V1 Knowledge Edges Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeDomainCreatePayload": {
        "properties": {
          "org_id": {
            "type": "string",
            "minLength": 1,
            "title": "Org Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "parent_domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Domain Id"
          },
          "sensitivity_default": {
            "type": "string",
            "title": "Sensitivity Default",
            "default": "internal"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "org_id",
          "name"
        ],
        "title": "KnowledgeDomainCreatePayload"
      },
      "KnowledgeEdgeCreatePayload": {
        "properties": {
          "org_id": {
            "type": "string",
            "minLength": 1,
            "title": "Org Id"
          },
          "from_node_id": {
            "type": "string",
            "minLength": 1,
            "title": "From Node Id"
          },
          "to_node_id": {
            "type": "string",
            "minLength": 1,
            "title": "To Node Id"
          },
          "relationship_type": {
            "type": "string",
            "minLength": 1,
            "title": "Relationship Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "default": "internal"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "org_id",
          "from_node_id",
          "to_node_id",
          "relationship_type"
        ],
        "title": "KnowledgeEdgeCreatePayload"
      },
      "KnowledgeNodeCreatePayload": {
        "properties": {
          "org_id": {
            "type": "string",
            "minLength": 1,
            "title": "Org Id"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "node_type": {
            "type": "string",
            "minLength": 1,
            "title": "Node Type"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "sensitivity": {
            "type": "string",
            "title": "Sensitivity",
            "default": "internal"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": ""
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "default": ""
          },
          "owner_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "org_id",
          "node_type",
          "title"
        ],
        "title": "KnowledgeNodeCreatePayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}